requirements/dev.txt
-r common.txt
coverage
sphinx
pylint
pep8
django-debug-toolbar
django-extensions
tox